refactor(ui): rename BadgeService ServiceIcon
