+++ libs/shared/ui/src/index.ts
@@ -83,7 +83,7 @@ export * from './lib/components/inputs/input-select/input-select'
 export * from './lib/components/table/table-row-deployment/table-row-deployment'
 export * from './lib/components/enable-box/enable-box'
 export * from './lib/components/property-card/property-card'
-export * from './lib/components/badge-service/badge-service'
+export * from './lib/components/service-icon/service-icon'
 export * from './lib/components/badge-deployment-order/badge-deployment-order'
 export * from './lib/components/date-picker/date-picker'
 export * from './lib/utils/toast'
